+++ Tests/Get-JiraConfigServer.Tests.ps1
@@ -1,7 +1,7 @@
 ﻿. $PSScriptRoot\Shared.ps1
 
 
-InModuleScope PSJira {
+InModuleScope JiraPS {
 
     [System.Diagnostics.CodeAnalysis.SuppressMessage('PSUseDeclaredVarsMoreThanAssigments', '', Scope='*', Target='SuppressImportModule')]
     $SuppressImportModule = $true