+++ Tools/psake.ps1
@@ -6,7 +6,7 @@ Properties {
     $ProjectRoot = $env:APPVEYOR_BUILD_FOLDER
 
     # This is the root of the module folder inside the project
-    $ModuleRoot = Join-Path -Path $ProjectRoot -ChildPath 'PSJira'
+    $ModuleRoot = Join-Path -Path $ProjectRoot -ChildPath 'JiraPS'
 
     $Commit = $env:APPVEYOR_REPO_COMMIT
     $CommitMessage = $env:APPVEYOR_REPO_COMMIT_MESSAGE
@@ -26,7 +26,7 @@ Task Init {
     $line
     Set-Location $ProjectRoot
     "Importing module"
-    Import-Module "$ProjectRoot\PSJira\PSJira.psd1"
+    Import-Module "$ProjectRoot\JiraPS\JiraPS.psd1"
     "`n"
 }
 
@@ -116,7 +116,7 @@ Task Deploy -Depends Build {
 
     if ($env:APPVEYOR_REPO_BRANCH -eq 'master' -and -not $env:APPVEYOR_PULL_REQUEST_NUMBER) {
         "Patching module manifest version with build number ($env:APPVEYOR_BUILD_NUMBER)"
-        Update-ModuleManifest -Path (Join-Path -Path $ModuleRoot -ChildPath 'PSJira.psd1') -BuildNumber $env:APPVEYOR_BUILD_NUMBER
+        Update-ModuleManifest -Path (Join-Path -Path $ModuleRoot -ChildPath 'JiraPS.psd1') -BuildNumber $env:APPVEYOR_BUILD_NUMBER
 
         $publishParams = @{
             Path = $ModuleRoot