+++ docs/changelog.rst
@@ -2,6 +2,18 @@
 Changelog
 =========
 
+2.0.0
+=====
+
+Release date: Jun 24, 2017
+
+Changes to the code module
+--------------------------
+
+* Move module to organization ``AtlassianPS``
+* Rename of the module to ``JiraPS`` [**breaking change**}
+* Rename of module's custom objects to ``JiraPS.*`` [**breaking change**]
+
 1.2.5
 =====